requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
django-auth-ldap==4.0.0
django-cors-headers==3.11.0
django-mssql==1.8
django-mssql-backend==2.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mod-wsgi==4.9.0
mssql-django==1.1.1
numpy==1.22.0
packaging==21.3
pandas==1.3.5
pandasql==0.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyodbc==4.0.32
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-ldap @ file:///E:/devs/python_ldap-3.4.0-cp310-cp310-win_amd64.whl
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
